w25q128jv: change programpage len to uint16_t
a page is 256 bytes, uint8 can only go to 255

+++ src/w25q128jv.c
@@ -69,7 +69,7 @@ void eraseChip() {
     CS_high(GPIOA, 4);
 }
 
-void programPage(uint32_t addr, uint8_t* tx, uint8_t* rx,  uint8_t len) {
+void programPage(uint32_t addr, uint8_t* tx, uint8_t* rx,  uint16_t len) {
     while(busy()) {}
     
     writeEnable();
